--- .ci/generate
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-make generate
